backend/requirements.txt
# System dependency: install ffmpeg and ensure it is available on PATH.
openai-whisper==20250625
torch==2.7.1
